lib: Map functions return long now

+++ src/libs/mysql/downtime.c
@@ -381,7 +381,7 @@ return tmprc;
   
 }
 
-int GetDowntimeMap(struct downtime * svcs, char * config, int orch_id) {
+long GetDowntimeMap(struct downtime * svcs, char * config, int orch_id) {
 	
 	MYSQL *mysql;
 	MYSQL_ROW  row;
@@ -392,7 +392,7 @@ int GetDowntimeMap(struct downtime * svcs, char * config, int orch_id) {
 	char * mysql_user = getConfigValue("mysql_user", config);
 	char * mysql_pw = getConfigValue("mysql_pw", config);
 	char * mysql_db = getConfigValue("mysql_db", config);
-	int i=0;
+	long i=0;
 	
 	char * sql, *where;
