+++ src/libs/mysql/servergroup.c
@@ -228,7 +228,7 @@ BARTLBY_SQL_PROTECTION_INIT;
     
 }
 
-int GetServerGroupMap(struct servergroup * svcs, char * config, int orch_id) {
+long GetServerGroupMap(struct servergroup * svcs, char * config, int orch_id) {
 	
 	MYSQL *mysql;
 	MYSQL_ROW  row;
@@ -239,7 +239,7 @@ int GetServerGroupMap(struct servergroup * svcs, char * config, int orch_id) {
 	char * mysql_user = getConfigValue("mysql_user", config);
 	char * mysql_pw = getConfigValue("mysql_pw", config);
 	char * mysql_db = getConfigValue("mysql_db", config);
-	int i=0;
+	long i=0;
 	
 		char * sql, *where;
