+++ src/libs/mysql/service.c
@@ -885,7 +885,7 @@ BARTLBY_SQL_PROTECTION_INIT;
 }
 
 
-int GetServiceMap(struct service * svcs, char * config, int orch_id) {
+long GetServiceMap(struct service * svcs, char * config, int orch_id) {
 	
 	MYSQL *mysql;
 	MYSQL_ROW  row;
@@ -895,7 +895,7 @@ int GetServiceMap(struct service * svcs, char * config, int orch_id) {
 	char * mysql_user = getConfigValue("mysql_user", config);
 	char * mysql_pw = getConfigValue("mysql_pw", config);
 	char * mysql_db = getConfigValue("mysql_db", config);
-	int i=0;
+	long i=0;
 
 	char * sql, *where;
