+++ src/libs/mysql/servicegroup.c
@@ -232,7 +232,7 @@ BARTLBY_SQL_PROTECTION_INIT;
     
 }
 
-int GetServiceGroupMap(struct servicegroup * svcs, char * config, int orch_id) {
+long GetServiceGroupMap(struct servicegroup * svcs, char * config, int orch_id) {
 	
 	MYSQL *mysql;
 	MYSQL_ROW  row;
@@ -243,7 +243,7 @@ int GetServiceGroupMap(struct servicegroup * svcs, char * config, int orch_id) {
 	char * mysql_user = getConfigValue("mysql_user", config);
 	char * mysql_pw = getConfigValue("mysql_pw", config);
 	char * mysql_db = getConfigValue("mysql_db", config);
-	int i=0;
+	long i=0;
 	
 	char * sql, *where;
