+++ src/libs/mysql/worker.c
@@ -531,7 +531,7 @@ BARTLBY_SQL_PROTECTION_INIT;
 	return rtc;	
 }	
 
-int GetWorkerMap(struct worker * svcs, char * config, int orch_id) {
+long GetWorkerMap(struct worker * svcs, char * config, int orch_id) {
 	
 	MYSQL *mysql;
 	MYSQL_ROW  row;
@@ -542,7 +542,7 @@ int GetWorkerMap(struct worker * svcs, char * config, int orch_id) {
 	char * mysql_user = getConfigValue("mysql_user", config);
 	char * mysql_pw = getConfigValue("mysql_pw", config);
 	char * mysql_db = getConfigValue("mysql_db", config);
-	int i=0;
+	long i=0;
 	
 char * sql, *where;
